[FileCommander] * Merge the archiver addons to one file * Merge ui.py to plugin.py * remove unneeded files
