+++ lib/python/Plugins/Extensions/FileCommander/Makefile.am
@@ -1,7 +1,5 @@
 installdir = $(pkglibdir)/python/Plugins/Extensions/FileCommander
 
-SUBDIRS = addons pic
-
 
 install_PYTHON = *.py
