--- lib/python/Plugins/Extensions/FileCommander/addons/Makefile.am
@@ -1,3 +1,0 @@
-installdir = $(pkglibdir)/python/Plugins/Extensions/FileCommander/addons
-
-install_PYTHON = *.py
