--- lib/python/Plugins/Extensions/FileCommander/pic/Makefile.am
@@ -1,3 +1,0 @@
-installdir = $(pkglibdir)/python/Plugins/Extensions/FileCommander/pic
-
-install_DATA = *.png
